+++ core/analytics/src/androidMain/AndroidManifest.xml
