+++ core/analytics/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/core/analytics/AnalyticsEvent.kt
